requirements.txt
selenium
pytest
pyyaml
Appium-Python-Client
